Pull classes out to their own file
